Fixes bug with debug=1 system setting
When you have debug=1 in system settings, manager pages with quip components are broken. This is due to the lines 128-131, which change the set error behaviour of your system (NEVER put out errors to HTML...). 
The revealing error was the access to config['debugUser'], which may not be set. This throws the error 

```
Notice: Undefined index: debugUser in /core/components/quip/model/quip/quip.class.php on line 132
```

This error was already reported at http://bugs.modx.com/issues/8467 and on other places, but was never corrected.